+++ CHANGES.rst
@@ -18,6 +18,7 @@
 - Add Check to _filter_object() if column contains only boolean (#110)
 - Add support of matrx and numpy.ndarray type of adata.X for predict_sex (#111)
 - Add method to get pd.DataFrame columns with list of regex (#90)
+- Added 'plot_pca_correlation' for plotting PCA correlation with obs/var columns (#118)
 
 ## Changes to ATAC notebooks (0.1.1.2 - 05-06-2023):   
 - Fixed assembling atac notebook 01